requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
python-dotenv>=1.0.1
pydantic>=2.8.2
pandas>=2.2.2
openpyxl>=3.1.5
python-docx>=1.1.2
python-pptx>=0.6.23
pypdf>=4.3.1
scikit-learn>=1.5.1
numpy>=1.26.4
SQLAlchemy>=2.0.36
psycopg[binary]>=3.2.3
pgvector>=0.3.6
openai>=1.58.1
langchain-openai>=0.2.14
